Removed unsupported submodule
WinGet-Merge is a work in progress feature that
is low priority and likely be on the back burner for a while
until other aspects of this repository are refined.

Granted having the option to combine two system's
suites of packages into a single shared list of packages
is ideal when trying to keep two systems' installed software
in sync within on another.