Added migration logic for winget.ignore
Whenever the ignore file is being accessed and no such file
exists, a check will be performed to determine if an instance
of this file can be migrated from a prior module install. If so,
this will be symlink'd or copied depending on the source's
own properties.

This commit also cleans up some log messages for consistency, and
resolves copy logic that was attempting to create a symlink instead
of copying the source file.