Removed admin check for symlink
Symlinks may not require admin if the user has been granted
privileges in the security settings policies. The upgrade process
for the module has replaced this check with a symlink creation test
as a means to skip the request if it fails.

There may be room to improve on this implementation; and in a
future update it might be desirable to give users the option to mute
warning messages.