Added busy indicators to checkpoint command
The main benefit of this is to properly provide terminal-level
busy indicators. Currently the winget command does not remain
busy through the execution of the command. This commit works around
this problem for consistency. A future update to winget may resolve
this, in which case, this commit may be reverted.

This commit also includes a delay for Merge-WinGetRestore. This
is mostly for aesthetics and give the user time to see the output from
the checkpoint command so it is not confused with some undesired
error output. There may be room for improvement on the handling here.