Fixed issue with cursor
After switching to the alternate screen buffer, the cursor does not
appear to hide via the use of [Console]::CursorVisible = $false.

Interestingly this issue can be resolved using the raw ANSI sequence
for this operation.

Along with this commit, a small tweak to the way the UI handles the
user input to exit the package details screen. The key presses will
now be cleared at the user types them.