Fixed issue undefined Show-JobProgress
During Update-WinGetEssentials, the module is removed and reloaded to apply
the latest version of the upgrade logic, however this does not import WinGet-Utils.psm1.
This resulted in the checkpoint and sync steps to encounter and error with Show-JobProgress
causing these minor steps to be skipped. This behavior is a bit odd and probably needs
further investigation and perhaps a reworked solution.